* Add skipTokenVerification requestmethod attribute * Add verifyPermission requestmethod attribute * RequestWorkers now check token and permissions depending on the selected requestmethod's attributes * Reflection yay * Move each requestmethod to its own file for readability * Refactoring yay * Remove some redundant code
